Add fonts and layout css
At first the fonts didnt work, apparently because of the format() I put, when i took that part off it all started working as it should.
There should be a background image / video for the site, either an image relted to food and dumplings or taiwanese street, or steam video looping in background to give more feeling. the images i have aren't seamless which doesn't look good in small screen. Will try to fix this later, for now just a black background.