Add about page load function and its css
added an about page load function, right now it appends to the page-content through the function itself, but later it will be appended through a load function (which will be used to load the other pages as well).
Also made the styling for the page, including small screen, maybe small changes later but all in all it looks fine.

Will also need to add a footer, added a first attempt in the constant load function, it's commented out because it's not finished.