Separate index.js page content creation to modules
The contact, about and menu creation are now on different modules, exporting only the final creation function, there is probably some more to clean up there. for now it works just fine. this also helps with the import of images, they will now be separated over the different modules, not a ton on indexjs.