Add clean:true to webpack config
The dist folder had some things that were unused, added clean to config so it will remove unused files. This also removed the index.html. Returned it but it's something to notice. Will need to add an htmlwebpackplugin with a template instead (or remove the clean:true as a hack).